move TerminationError to spiceerrors package
